src/main/java/ehub/com/pages/S04createNewOnshorePolicyWithTermsAndScheduleItemsPage.java
package ehub.com.pages;

import org.openqa.selenium.WebDriver;
import org.openqa.selenium.support.PageFactory;

import ehub.com.constants.AppConstants;

public class S04createNewOnshorePolicyWithTermsAndScheduleItemsPage {
	
	//private WebDriver driver;
	CommonLocatorsAndMethod common;

	public S04createNewOnshorePolicyWithTermsAndScheduleItemsPage(WebDriver driver) {
		//this.driver = driver;
		PageFactory.initElements(driver, this);
		common = new CommonLocatorsAndMethod(driver);
	}

	public void login() {
		common.login(AppConstants.username, AppConstants.password);
	}

	public String createNewPolicy(String policyNumber) {
		 String createdPolicyVerify =common.createNewPolicyWithClassNameAndCoverageRule(policyNumber, AppConstants.policy_Assured + "04", 2, AppConstants.className_two_class_With_Same_Name);
		 return createdPolicyVerify;
	}
	
	public String addNewLimitGroup(String groupName) {
		String actual=common.addNewLimitGroup(groupName, AppConstants.country_Australia);
		return actual;
	}
	
	public String addNewLimit(String validationMessage) {
		String actual=common.addNewLimit(validationMessage);
		return actual;
	}
	
	public String addSchedule(String validationMessage) {
		String actual=common.addSingleScheduleGeneric(AppConstants.class_Cyber_Fraction_Value, validationMessage);
		return actual;
	}

}
